refactor: route core sampling through platform traits
Move realtime GPU and live sensor sampling behind the core platform trait layer so collector sampling stays OS-agnostic orchestration.

Add typed sensor availability results for unsupported and unavailable live sensor paths, preserving External Component Guidance behavior for attempted Windows PawnIO collection.

Keep Windows-only GPU BDF name resolution private to the Windows GPU platform module, and keep Linux lspci lookup off the async executor during sampling.

Validation: cargo fmt --all; cargo test -p hardviz-core --locked --offline; git diff --check.
