+++ core/src/platform/linux/mod.rs
@@ -4,7 +4,7 @@ use crate::models::hardware::{
 };
 use crate::platform::traits::{
   GpuPlatform, MemoryPlatform, MotherboardPlatform, NetworkPlatform, Platform,
-  ProcessElevationPlatform, SuperIoPlatform,
+  ProcessElevationPlatform, SensorPlatform, SuperIoPlatform,
 };
 use std::future::Future;
 use std::pin::Pin;
@@ -80,6 +80,12 @@ impl GpuPlatform for LinuxPlatform {
   {
     Box::pin(async { Ok(None) })
   }
+
+  fn sample_gpus(
+    &self,
+  ) -> Pin<Box<dyn Future<Output = Vec<crate::models::GpuSample>> + Send + '_>> {
+    Box::pin(gpu::sample_gpus())
+  }
 }
 
 impl NetworkPlatform for LinuxPlatform {
@@ -107,6 +113,20 @@ impl MotherboardPlatform for LinuxPlatform {
 impl SuperIoPlatform for LinuxPlatform {
   fn get_super_io_chip_id_diagnostics(&self) -> SuperIoChipIdDiagnostics {
     SuperIoChipIdDiagnostics::unsupported_platform()
+  }
+}
+
+impl SensorPlatform for LinuxPlatform {
+  fn sample_temperatures(&self) -> crate::models::TemperatureSample {
+    crate::models::TemperatureSample::unsupported(
+      "CPU and named sensor temperature sampling is not implemented for LinuxPlatform",
+    )
+  }
+
+  fn sample_motherboard_sensors(&self) -> crate::models::MotherboardSensorCollection {
+    crate::models::MotherboardSensorCollection::unsupported(
+      "Motherboard sensor sampling is available on Windows only",
+    )
   }
 }
