+++ core/src/platform/macos/mod.rs
@@ -4,7 +4,7 @@ use crate::models::hardware::{
 };
 use crate::platform::traits::{
   GpuPlatform, MemoryPlatform, MotherboardPlatform, NetworkPlatform, Platform,
-  ProcessElevationPlatform, SuperIoPlatform,
+  ProcessElevationPlatform, SensorPlatform, SuperIoPlatform,
 };
 use std::future::Future;
 use std::pin::Pin;
@@ -79,6 +79,12 @@ impl GpuPlatform for MacOSPlatform {
   {
     Box::pin(async { gpu::get_gpu_memory_usage().await })
   }
+
+  fn sample_gpus(
+    &self,
+  ) -> Pin<Box<dyn Future<Output = Vec<crate::models::GpuSample>> + Send + '_>> {
+    Box::pin(gpu::sample_gpus())
+  }
 }
 
 impl NetworkPlatform for MacOSPlatform {
@@ -104,6 +110,20 @@ impl MotherboardPlatform for MacOSPlatform {
 impl SuperIoPlatform for MacOSPlatform {
   fn get_super_io_chip_id_diagnostics(&self) -> SuperIoChipIdDiagnostics {
     SuperIoChipIdDiagnostics::unsupported_platform()
+  }
+}
+
+impl SensorPlatform for MacOSPlatform {
+  fn sample_temperatures(&self) -> crate::models::TemperatureSample {
+    crate::models::TemperatureSample::unsupported(
+      "CPU and named sensor temperature sampling is not implemented for MacOSPlatform",
+    )
+  }
+
+  fn sample_motherboard_sensors(&self) -> crate::models::MotherboardSensorCollection {
+    crate::models::MotherboardSensorCollection::unsupported(
+      "Motherboard sensor sampling is available on Windows only",
+    )
   }
 }
